Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/at-property-stylesheets-expected.txt


FAIL @property detected when stylesheet appears assert_equals: expected "1px" but got ""
FAIL @property removal detected when last @property rule disappears assert_equals: expected "1px" but got ""
FAIL @property detected in second stylesheet assert_equals: expected "2px" but got ""
FAIL @property removal detected with removal of second stylesheet assert_equals: expected "2px" but got ""
FAIL @property removal detected with removal of first stylesheet assert_equals: expected "1px" but got ""
